calicore/requirements.txt
streamlit
google-generativeai
python-dotenv
fpdf
python-docx
